+++ spec/unit/cluster_config_spec.rb
@@ -1,19 +1,9 @@
 require 'kube_deploy_tools/cluster_config'
 
 describe 'cluster config' do
-  it "collects environmental Git information to create an appropriately long tag value" do
-    ENV['GIT_COMMIT'] = '0981b78141b123965e7380d6021f7a9b76426290'
-    ENV['GIT_BRANCH'] = 'my-testing-branch'
-    ENV.delete('BUILD_ID')
-
+  it "creates a valid Docker tag" do
     tag = KubeDeployTools.tag_from_local_env
-    expect(tag).to eq('my-testing-branch-0981b78-dev')
-
-    ENV['GIT_BRANCH'] = 'my testing! branch with some weird# characters, and also overflowing'
-    ENV['BUILD_ID'] = '12345'
-    tag = KubeDeployTools.tag_from_local_env
-    expect(tag).to eq('my_testing__branch_with_some_weird__characters__a-0981b78-12345')
-    expect(tag.size).to be <= 63
+    expect(tag).to match(/[\w][\w.-]{0,127}/)
   end
 
   # Mock shellrunner